+++ src/include/index/vamana_metadata.h
@@ -117,7 +117,7 @@ class vamana_index_metadata
 
   void clear_history_impl(uint64_t timestamp) {
     std::vector<num_edges_history_type> new_num_edges_history;
-    for (int i = 0; i < ingestion_timestamps_.size(); i++) {
+    for (uint64_t i = 0; i < ingestion_timestamps_.size(); i++) {
       auto ingestion_timestamp = ingestion_timestamps_[i];
       if (ingestion_timestamp > timestamp) {
         new_num_edges_history.push_back(num_edges_history_[i]);
